Fix the resizing issue when enabling manual sizing.
When enabling manual sizing width and height of the cell is changed to the last saved by the user. Instead, the expected behavior is so width and height should be the same as before enabling manual sizing. It requires persisting current width and height of cells, but it triggers an additional update.